feat: Change API to expose only an Asset construct, and not an entire LayerVersion (#4)
The reason for this change is that keeping any changes in the LayerVersion API in sync between this library and aws-cdk-lib is not maintainable. Instead, this library will just expose an `Asset` construct that can be used in `aws-cdk-lib` to create the LayerVersion. 

Usage in aws-cdk-lib will look like this:

```
export class AwsCliLayer extends lambda.LayerVersion {
  private static assetPackage: any;

  constructor(scope: Construct, id: string) {
    AwsCliLayer.assetPackage = // Dynamic require logic goes here...

    super(scope, id, {
      code: lambda.Code.fromAssetConstruct(new AwsCliLayer.assetPackage.AwsCliAsset(scope, `${id}-asset`),
      description: '/opt/awscli/aws',
    });
  }
}
```

Any ideas for better naming in this PR are greatly appreciated!

I tested this locally by `npm link`-ing to aws-cdk-lib and it works.