chore: upgrade workflow for aws-cli Dockerfile dependency (#83)
I developed this in my fork; the result of this workflow will be a PR that looks like [this](https://github.com/kaizencc/awscdk-asset-awscli/pull/2).

Most of the github workflow is ripped from the `upgrade-awscli-v2-main` workflow, except that this replaces the projen step:

```yml
      - name: Install Python Script Dependencies
        run: |
          sudo apt update
          sudo apt install -y jq
          pip install semver
      - name: Check for awscli version upgrades
        run: python3 .github/scripts/upgrade-awscli-version.py
``` 

Upon merging, this should immediately introduce a PR that upgrades the aws-cli dep to 2.9.0.