chore: backport v2 upgrade workflow to default branch so it shows up in the actions tab (#146)
In "this totally makes no sense yet is how the world works" news, workflows don't exist at all unless they exist in the default branch. And then, you can specify which branch to run it off of. We want to run this workflow in `awscli-v2/main`, but since it doesn't exist in `awscli-v1/main`, it doesn't exist at all. This PR should change that.

This file exists today in `awscli-v2/main`, so its just a backport. The file was introduced in https://github.com/cdklabs/awscdk-asset-awscli/pull/83