fix: update AWS CLI to 2.2.201 (#645)
This library has not been released since July. This could be a potential source of a `urllib3` security notification. At the least, releasing it makes sense so that we can test if this indeed resolved the issue.

We only release for `feat` and `fix` and all upgrades are `chores`, hence no releases since July.