chore: fix the script upgrading the aws cli version
This was previously using sed. Which is fine, but sed behaves differently on Linux and macOs. So I wasted a bunch of time trying to understand why it's working locally but not in GitHub Action. Having scripts behave differently in different environments isn't great. So instead of fixing the sed command, this is now using python only.